Moved takeoff-related segments in takeoff package.
