+++ src/fastoad/models/performances/mission/segments/takeoff/__init__.py
@@ -1,3 +1,8 @@
+"""
+Classes for simulating takeoff-related flight segments.
+
+Be sure to import this package before interpreting a mission input file.
+"""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
 #  Copyright (C) 2023 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
@@ -10,3 +15,10 @@
 #  GNU General Public License for more details.
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+# pylint: disable=unused-import
+# flake8: noqa
+
+# With these imports, importing only the current package ensures to have all
+# these segments available when interpreting a mission input file
+from . import end_of_takeoff, rotation, takeoff
