+++ src/fastoad/models/performances/mission/segments/takeoff/rotation.py
@@ -1,6 +1,6 @@
 """Classes for acceleration/deceleration segments."""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
+#  Copyright (C) 2023 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -16,11 +16,11 @@ import logging
 from dataclasses import dataclass
 from typing import List
 
-from numpy import pi, sin, cos
+from numpy import cos, pi, sin
 from scipy.constants import g
 
 from fastoad.model_base import FlightPoint
-from .base import GroundSegment
+from fastoad.models.performances.mission.segments.base import GroundSegment
 
 _LOGGER = logging.getLogger(__name__)  # Logger for this module
