+++ src/fastoad/models/performances/mission/segments/takeoff/takeoff.py
@@ -18,10 +18,10 @@ from fastoad.constants import EngineSetting
 from fastoad.model_base import FlightPoint
 from fastoad.model_base.datacls import MANDATORY_FIELD
 from fastoad.model_base.propulsion import IPropulsion
+from .end_of_takeoff import EndOfTakoffSegment
+from .rotation import RotationSegment
 from ..base import DEFAULT_TIME_STEP, RegisteredSegment
-from ..end_of_takeoff import EndOfTakoffSegment
 from ..ground_speed_change import GroundSpeedChangeSegment
-from ..rotation import RotationSegment
 from ...base import FlightSequence
 from ...polar import Polar
 from ...polar_modifier import AbstractPolarModifier, UnchangedPolar
